add use case context

+++ src/modules/queue/useCases/DisconnectUseCase/DisconnectUseCase.ts
@@ -1,4 +1,4 @@
-import { UseCase } from "@core";
+import { IUseCaseContext, UseCase } from "@core";
 import { IQueueRepository } from "@modules/queue";
 import Joi from "joi";
 import { inject, injectable } from "tsyringe";
@@ -19,11 +19,14 @@ export class DisconnectUseCase extends UseCase<Params, Response> {
 		super();
 	}
 
-	public async run(params: Params): Promise<Response> {
+	public async run(params: Params, { userId }: IUseCaseContext): Promise<Response> {
 		const { guildId } = params;
 
 		const queue = this.queueRepository.get(guildId);
 		if (!queue) throw new Error("Queue not found");
+		if (!queue.voiceChannel.members.find((m) => m.id === userId)) {
+			throw new Error("User not in voice channel");
+		}
 
 		queue.stop();
 		this.queueRepository.delete(guildId);
